Adding explicit Xi values to MultiVector creation
We've always had explicit Xi values when defining a Xi or Pair but this
was never hooked into in the MultiVector API. The new syntax is to allow
for string initiated MultiVectors to contain an explicit Xi value
following each Alpha contained within square brackets. The corresponding
change has also been made to the calculation file parsing logic and an
example file has been added.