Fixing bug with custom differentials
So far this hasn't caused a problem as we have been explicitly passing
the config into all differential calls defined under a custom config but
the repr of such differentials was broken. In looking into why I
realised that we always defaulted to the master config for doing the
differential NOT the one passed at init. This is now fixed.

This commit also adds the ability to create a multivector inside of an
ARContext which simplifies the construction of custom multivectors.