Application of Zet/Nat to simplify things and renaming 4set/fourset to zet
This is a little bit messy but the Zet/Nat functions have now replaced
multiple helper properties on the ARConfig object as they are invariant
with respect to metric and allowed. In doing this it seemed prudent to
also clear up the old terminology of `4set/fourset/four_set` and have
everything in terms of `Zet`. Test suite is passing now but there may be
some subtle things knocking around that I've missed...Will need a proper
review before making the next release!