Fixing comparison of Xi/Terms by partial derivative
Given that d^2/dxdy == d^2/dydx we should simply sort the partial
derivatives on a Xi / Term each time they are updated and then use the
existing logic to cancel terms that have matching partials (now that
they are in a standard form order).

We also stop auto cancelling terms on creation of MultiVectors in order
to allow users to inspect the raw terms that are generated BEFORE
cancelling them manually using the renamed 'cancel_terms()' method on
MultiVector instances.